server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
grpcio==1.73.0
grpcio-tools==1.73.0
protobuf==6.31.1
Jinja2==3.1.6
matplotlib==3.10.3
